frontend

+++ src/components/dashboard/Files/Files.jsx
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from 'react'
 import './Files.css';
 import { useNavigate } from 'react-router-dom';
 import { getAllFiles , deleteFile } from '../../../Services/FileService'; 
+import Swal from 'sweetalert2';
 
 const Files = () => {
 
@@ -26,11 +27,35 @@ const Files = () => {
   }, []);
 
   const handleDelete = async (id) => {
-    try {
-      await deleteFile(id);
-      setFiles(files.filter((file) => file._id !== id));
-    } catch (error) {
-      console.error('Erreur lors de la suppression du fichier :', error);
+    const result = await Swal.fire({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this file!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#3085d6",
+      confirmButtonText: "Yes, delete!",
+      cancelButtonText: "Cancel"
+    });
+
+    if (result.isConfirmed) {
+      try {
+        await deleteFile(id);
+        setFiles(files.filter((file) => file._id !== id));
+
+        Swal.fire(
+          "Deleted!",
+          "The file has been successfully deleted.",
+          "success"
+        );
+      } catch (error) {
+        Swal.fire(
+          "Error!",
+          "Failed to delete the file.",
+          "error"
+        );
+        console.error("Error deleting file:", error);
+      }
     }
   };
 
@@ -72,11 +97,13 @@ const Files = () => {
               <div className="file-name">{file.name}</div>
               <div className="file-details">
                 <span>Updated At : {file.updatedAt}</span>
+                <span>Created By: {file.userId?.name || 'User'}</span>
+                <div className="file-actions">
+                <button className='btnEdit' onClick={() => handleEdit(file)}>Edit</button>
+                <button className='btnDelete' onClick={() => handleDelete(file._id)}>Delete</button>
               </div>
-              <div className="file-actions">
-                <button onClick={() => handleEdit(file)}>Edit</button>
-                <button onClick={() => handleDelete(file._id)}>Delete</button>
               </div>
+              
             </div>
           ))
         ) : (
